Remove redundant test (which used DSA)
