+++ close_test.go
@@ -24,51 +24,12 @@ package crypto11
 import (
 	"crypto/dsa"
 	"crypto/elliptic"
-	"math/rand"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/assert"
 
 	"github.com/stretchr/testify/require"
 )
-
-func TestClose(t *testing.T) {
-	// Verify that close and re-open works.
-
-	ctx, err := ConfigureFromFile("config")
-	require.NoError(t, err)
-
-	id := randomBytes()
-	_, err = ctx.GenerateRSAKeyPair(id, rsaSize)
-	if err != nil {
-		_ = ctx.Close()
-		t.Fatal(err)
-	}
-
-	require.NoError(t, ctx.Close())
-
-	ctx, err = ConfigureFromFile("config")
-	require.NoError(t, err)
-
-	key2, err := ctx.FindKeyPair(id, nil)
-	require.NoError(t, err)
-
-	testRsaSigning(t, key2, false)
-	_ = key2.Delete()
-	require.NoError(t, ctx.Close())
-}
-
-// randomBytes returns 32 random bytes.
-func randomBytes() []byte {
-	result := make([]byte, 32)
-	rand.Read(result)
-	return result
-}
-
-func init() {
-	rand.Seed(time.Now().UnixNano())
-}
 
 func TestErrorAfterClosed(t *testing.T) {
 	ctx, err := ConfigureFromFile("config")
